+++ CHANGELOG.rst
@@ -2,9 +2,13 @@ Changelog
 =========
 
 TBD
----
+------------------
+* #650 Fixed space encoding in base string URI used in the signature base string.
+* #652: Fixed OIDC /token response which wrongly returned "&state=None"
 
-* Fixed space encoding in base string URI used in the signature base string.
+3.0.1 (2019-01-24)
+------------------
+* Fixed OAuth2.0 regression introduced in 3.0.0: Revocation with Basic auth no longer possible #644
 
 3.0.0 (2019-01-01)
 ------------------